fix(665): clean
